+++ james/james-core/CHANGES.md
@@ -11,6 +11,7 @@ Version 0.2 (under development)
  - Removed SubsetProblem interface and renamed SubsetProblemWithData to SubsetProblem.
  - Removed MinMaxObjective.
  - Removed EmptySearchListener and EmptyLocalSearchListener. Default empty implementations of all callbacks are now directly provided in the corresponding interfaces.
+ - Refactored abstract Solution class.
  - Reorganized package structure.
  - Moved to SLF4J API 1.7.7.
  - Moved to Java 8.